feat(100-02): create run_shap_analysis CLI and run SHAP analysis
- run_shap_analysis.py: argparse CLI with --tf, --venue-id, --max-samples,
  --top-k, --top-k-features, --update-yaml, --report-path, --model-path flags
- Loads or trains CrossSectionalRanker via train_full() + pickle cache
- Runs compute_shap_values() + compute_interaction_values() on full feature matrix
- Prints top interaction pairs and top SHAP features to stdout
- Writes markdown report to reports/ml/shap_interaction_report.md
- --update-yaml writes 'interactions' key to configs/feature_selection.yaml
- Logs experiment to ml_experiments via ExperimentTracker

Rule 1 bugfix in ranker.py: train_full() lacked astype(float) cast on X;
object-dtype columns with Python None caused np.nanmedian TypeError.

Rule 1 bugfixes in shap_analysis.py:
- LIKE clause uses %% (not %) in SQLAlchemy text() to avoid psycopg2 format error
- dim_feature_selection uses 'rationale' column (not 'notes')

Results (tf=1D, 7 assets, 126 features, 23176 rows):
  Top interaction pairs:
    1. bb_ma_20 x close_fracdiff     (strength=0.1192)
    2. close_fracdiff x ret_is_outlier (strength=0.0017)
    3. bb_ma_20 x ret_is_outlier     (strength=0.0017)
  Top SHAP features:
    1. bb_ma_20        (mean|SHAP|=0.1411)
    2. close_fracdiff  (mean|SHAP|=0.1187)
    3. ret_is_outlier  (mean|SHAP|=0.0033)

+++ src/ta_lab2/ml/ranker.py
@@ -625,7 +625,10 @@ class CrossSectionalRanker:
         feature_cols = self._get_feature_cols(df)
         self.feature_names_ = feature_cols
 
-        X = df[feature_cols].values
+        # Cast to float64 so Python None values become np.nan (not object dtype).
+        # Without this, np.nanmedian raises TypeError when object columns
+        # contain Python None instead of np.nan.
+        X = df[feature_cols].astype(float).values
         y = self._build_rank_target(df).values
         group = self._build_group_array(df["ts"])
